requirements.txt
-e .
aws_cdk.aws_s3==1.70.0
aws_cdk.aws_glue==1.70.0
aws_cdk.aws_iam==1.70.0
aws_cdk.aws_athena==1.70.0
aws_cdk.aws_kinesisfirehose==1.70.0